requirements.txt
streamlit
openai
cohere
google-generativeai
beautifulsoup4
requests
PyMuPDF
